requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
pandas==2.1.4
numpy==1.26.2

# API and networking
httpx==0.26.0
aiohttp==3.9.1
aioresponses==0.7.6
tenacity==8.2.3
pytz==2023.3
eodhd==1.0.32
anthropic>=0.20.0

# Notification systems
twilio==8.11.0

# Scheduling and async
apscheduler==3.10.4
asyncio==3.4.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development tools
black==23.12.1
ruff==0.1.11
mypy==1.8.0
pre-commit==3.6.0

# Logging and monitoring
structlog==24.1.0
python-json-logger==2.0.7